refactor(chessboard): extract the game machine from the screen
GameScreen had grown into a god component owning five unrelated
machines. Pull them apart:

- use-chess-game.ts — the whole game behaviour behind one hook: replay
  loop with pause/resume and generation-based cancellation, per-move
  state + shared-value writes, haptics, the chess-clock ticker and the
  checkmate-aura trigger. The screen is pure layout now (~200 lines).
- haptics.ts — the MATE_PATTERN composition and the per-move tick map.
- components/board.tsx, play-pause-button.tsx, status-caption.tsx —
  each leaf owns its styles.
- constants: FATAL_ATTRACTION is typed [Square, Square][] (kills the
  `as any` casts on move()), and the magic replay timings live in a
  named REPLAY object with the reasoning documented.
- utils: statusFor() replaces the nested status ternary.

Behaviour-preserving — replay, pause, clocks, mate aura and recap
verified on the simulator.
